be_sock: do not call writecb if the EV_WRITE is disabled (during connecting)
Most of time this is not a problem, since EV_WRITE is disabled just
after connection succeed (if it was not enabled prior to the
bufferevent_socket_connect() call), however if the event_base_loop() and
bufferevent_socket_connect() called from different threads then
bufferevent_writecb() can be called before "connecting" flag will be
installed and it will call user specified callback (even if it was not
enabled).

And even installing "connecting" flag at the beginning of the
bufferevent_socket_connect() will not fix the problem, because if to the
moment of calling bufferevent_socket_connect() EV_WRITE was enabled then
the EV_WRITE will not be disabled and the writecb will be called in
parallel which is also bad (and in case of http it will run
evhttp_request_dispatch() mulitle time for the same request, and
evcon->state will be incorrect after).

You can trigger this issue via http layer, by running event_base_loop()
and evhttp_make_request() in different threads, like in the stacktrace
below (also note, that this stacktrace gathered without
BEV_OPT_THREADSAFE, which is mandratory here):

T1:
- be_socket_setfd (fd=1320)
- be_socket_ctrl
- bufferevent_setfd
- bufferevent_socket_connect
- bufferevent_connect_getaddrinfo_cb
- evdns_getaddrinfo
- evutil_getaddrinfo_async_
- bufferevent_socket_connect_hostname
- evhttp_connection_connect_ # can  have bufferevent_setfd()
- evhttp_make_request

T2:
- bufferevent_writecb (fd=1320)
- event_persist_closure
- event_process_active_single_queue
- event_process_active
- event_base_loop

Fixes: #733